+++ src/autoanki/Dictionary/CEDictionary.py
@@ -1,10 +1,47 @@
 from .Dictionary import Dictionary
 from os import path
-import re
 import logging
-from pprint import pprint
+import pinyin
+
+import jieba
+import jieba.posseg as pseg
+import chinese_converter
 
 PATH_TO_FILE = path.join(path.dirname(__file__), 'cedict_ts.u8')
+
+#   Use the lookup table here: https://github.com/fxsjy/jieba
+POS_LUT = {
+    'n':'noun',
+    'nt':'noun',
+    'nrt':'noun',
+    'ORG':'noun',
+    'nr':'noun',
+    'PER':'noun',
+    'ns':'noun',
+    'nw':'noun',
+    'nz':'noun',
+    'f':'noun',
+    's':'noun',
+    'an':'noun',
+    't':'time',
+    'TIME':'time',
+    'p':'adposition',
+    'r':'pronoun',
+    'q':'quantifier',
+    'm':'quantifier',
+    'v':'verb',
+    'vd':'verb',
+    'vn':'verb',
+    'a':'adjective',
+    'ad':'adverb',
+    'd':'adverb',
+    'c':'conjunction',
+    'u':'particle',
+    'vn':'other',
+    'xc':'other',
+    'x':'other',
+    'w':'puncuation',
+}
 
 class CEDictionary(Dictionary):
 
@@ -14,12 +51,14 @@ class CEDictionary(Dictionary):
         """
         self.logger = logging.getLogger('autoanki.cedict')
         self.logger.setLevel(debug_level)
+        self.logger.info("Loading Chinese Dictionary")
         self.dict = self._parse_file()
+        self.logger.info("Done!")
         
         pass
 
     def _parse_file(self) -> dict[str, dict]:
-        self.logger.info("Parsing dict file...")
+        self.logger.info("Parsing file...")
         if not path.isfile(PATH_TO_FILE):
             self.logger.critical("Could not open dictionary file")
 
@@ -27,7 +66,7 @@ class CEDictionary(Dictionary):
         #   e.g. 哈利·波特
         #   this does not currently match
         definitions = {}
-        last_word = "~~"
+        last_word = "~"
         current_word = ""
 
         with open(PATH_TO_FILE, "r") as dict_file:
@@ -37,36 +76,41 @@ class CEDictionary(Dictionary):
                 parts = line.split(" ")
                 trad_word = parts[0]
                 current_word = parts[1]
-                
-                m = re.compile(r"\[[^]]*\]")
-                match = m.findall(line)
-                # self.logger.debug(match)
-                pinyin_numbers = match[0]
 
                 definition = line[line.find("]")+2:]
 
                 if current_word != last_word:
                     # self.logger.debug(f"Creating entry: {current_word}")
-                    definitions[current_word] = {'trad_word': "", 'pinyin_numbers':"", 'definition':""}
+                    definitions[current_word] = {
+                        'trad_word': "", 
+                        'pinyin_numbers':"", 
+                        'definition':"", 
+                        'pinyin':""
+                    }
                 
-                # self.logger.debug(f"Word: {current_word}")
-                # self.logger.debug(f"Trad: {trad_word}")
-                # self.logger.debug(f"PinYin: {pinyin_numbers}")
-                # self.logger.debug(f"DEF: {definition}")
-                definitions[current_word]['trad_word'] = trad_word
-                definitions[current_word]['pinyin_numbers'] += '<br>' + pinyin_numbers 
-                definitions[current_word]['definition'] += '<br>' + definition 
+                # These only need to be done the first time the word is seen
+                if last_word != current_word:
+                    definitions[current_word]['trad_word'] = trad_word
+                    definitions[current_word]['pinyin'] = pinyin.get(current_word)
+                    definitions[current_word]['pinyin_numbers'] = pinyin.get(current_word, format="numerical")
 
+
+                definitions[current_word]['definition'] += '<br>' + definition 
                 last_word = current_word
 
-        self.logger.info("Done!")
-        for key, value in definitions.items():
+        for key, _ in definitions.items():
             definitions[key]['definition'].strip('\n')
-            # print(key, '->', value)
         return definitions
 
 
     def find_word(self, word: str) -> None | list[str]:
+        """
+        Find a word in the dictionary. This can be simplified, or traditional
+        `return` Paramaters that get passed to the database
+        """
+        
+        # Convert the word to simplified if needed
+        word = chinese_converter.to_simplified(word)
 
         if word in self.dict:
             """
@@ -81,9 +125,17 @@ class CEDictionary(Dictionary):
             word = params[8]
             """
             params = ["", "", "", "", "", "", "", "", ""]
+
+            slice = next(pseg.cut(word))
+            part_of_speech = POS_LUT.get(slice.flag)
+            if part_of_speech != None:
+                params[1] = part_of_speech
+
             params[0] = self.dict[word]['trad_word'] 
+            params[2] = self.dict[word]['pinyin'] 
             params[3] = self.dict[word]['pinyin_numbers'] 
             params[7] = self.dict[word]['definition'] 
+            self.logger.debug(f"Word: {word}")
             params[8] = word
             return params
         else: